refactor(logs): hoist log-text helpers to utils, simplify CLI
Polish pass on the sandbox-logs branch. Mostly dedup and cleanup.

- Add `microsandbox_utils::log_text` with `strip_ansi`,
  `split_leading_timestamp`, `base64_decode`. Both the SDK reader
  (`microsandbox::sandbox::logs`) and the CLI renderer
  (`msb logs`) consume from there instead of carrying parallel copies.
- Drop dead `EffectiveSource` enum + match in `msb logs`; the result
  was computed and immediately discarded.
- `follow_loop` is now `async`, uses `tokio::time::sleep` instead of
  `std::thread::sleep` (was blocking a tokio worker), hoists the
  `--grep` regex out of the polling tick, and collapses three
  per-tick `metadata`+`inode` syscalls into one.
- Inline `resolve_log_dir` (was a one-liner `Ok(...)` wrapper).
- Boot-error JSON-Lines entry now follows the documented schema:
  `d` is a string, no extra sibling field.
- Relay tap removes registry entries on any FLAG_TERMINAL frame
  (covers ExecExited and ExecFailed uniformly) so abnormal client
  disconnects no longer leak `SessionInfo`.
- Trim a couple of narrative comments that paraphrased the code.
