+++ crates/microsandbox/lib/sandbox/logs.rs
@@ -12,6 +12,7 @@ use std::path::{Path, PathBuf};
 
 use bytes::Bytes;
 use chrono::{DateTime, Utc};
+use microsandbox_utils::log_text::{base64_decode, split_leading_timestamp, strip_ansi};
 use serde::Deserialize;
 
 use crate::{MicrosandboxError, MicrosandboxResult};
@@ -268,61 +269,6 @@ fn append_text_log_as_system(path: &Path, out: &mut Vec<LogEntry>) {
     }
 }
 
-/// Strip ANSI escape sequences (CSI, OSC, two-byte C1).
-///
-/// Hand-rolled state machine. Avoids pulling the `regex` crate just
-/// for one fixed pattern. Handles:
-///
-/// - `\x1b[…<final>` — CSI (SGR colors, cursor moves). Final byte is
-///   in `0x40..=0x7e`.
-/// - `\x1b]…\x07` and `\x1b]…\x1b\\` — OSC (terminated by BEL or ST).
-/// - `\x1b<X>` for X in `0x40..=0x5f` — two-byte C1 controls.
-fn strip_ansi(input: &str) -> String {
-    let mut out = String::with_capacity(input.len());
-    let mut chars = input.chars().peekable();
-    while let Some(c) = chars.next() {
-        if c != '\x1b' {
-            out.push(c);
-            continue;
-        }
-        match chars.next() {
-            Some('[') => {
-                // CSI: skip until final byte in 0x40..=0x7e.
-                for c in chars.by_ref() {
-                    if matches!(c, '\x40'..='\x7e') {
-                        break;
-                    }
-                }
-            }
-            Some(']') => {
-                // OSC: skip until BEL or ESC '\'.
-                while let Some(c) = chars.next() {
-                    if c == '\x07' {
-                        break;
-                    }
-                    if c == '\x1b' && chars.peek() == Some(&'\\') {
-                        chars.next();
-                        break;
-                    }
-                }
-            }
-            // Two-byte C1 (or stray ESC) — drop the next char.
-            Some(_) => {}
-            None => break,
-        }
-    }
-    out
-}
-
-fn split_leading_timestamp(line: &str) -> Option<(&str, &str)> {
-    let (first, rest) = line.split_once(char::is_whitespace)?;
-    if first.len() >= 20 && first.ends_with('Z') {
-        Some((first, rest))
-    } else {
-        None
-    }
-}
-
 fn apply_filters(entries: &mut Vec<LogEntry>, opts: &LogOptions) {
     if let Some(s) = opts.since {
         entries.retain(|e| e.timestamp >= s);
@@ -336,43 +282,6 @@ fn apply_filters(entries: &mut Vec<LogEntry>, opts: &LogOptions) {
         let drop = entries.len() - n;
         entries.drain(0..drop);
     }
-}
-
-fn base64_decode(s: &str) -> Option<Vec<u8>> {
-    static TABLE: &[u8] = b"ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789+/";
-    let bytes = s.trim().as_bytes();
-    if bytes.is_empty() {
-        return Some(Vec::new());
-    }
-    if !bytes.len().is_multiple_of(4) {
-        return None;
-    }
-    let mut out = Vec::with_capacity(bytes.len() / 4 * 3);
-    for chunk in bytes.chunks(4) {
-        let mut vals = [0u8; 4];
-        let mut pad = 0usize;
-        for (i, &b) in chunk.iter().enumerate() {
-            if b == b'=' {
-                pad += 1;
-                vals[i] = 0;
-            } else {
-                let idx = TABLE.iter().position(|&t| t == b)?;
-                vals[i] = idx as u8;
-            }
-        }
-        let n = ((vals[0] as u32) << 18)
-            | ((vals[1] as u32) << 12)
-            | ((vals[2] as u32) << 6)
-            | (vals[3] as u32);
-        out.push(((n >> 16) & 0xff) as u8);
-        if pad < 2 {
-            out.push(((n >> 8) & 0xff) as u8);
-        }
-        if pad < 1 {
-            out.push((n & 0xff) as u8);
-        }
-    }
-    Some(out)
 }
 
 //--------------------------------------------------------------------------------------------------
