+++ crates/utils/lib/lib.rs
@@ -1,5 +1,6 @@
 //! Shared constants and utilities for the microsandbox project.
 
+pub mod log_text;
 pub mod size;
 pub mod ttl_reverse_index;
 pub mod wake_pipe;
